james_library-main/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rain-lab"
version = "0.2.0"
description = "R.A.I.N. Lab — Recursive Architecture of Intelligent Nexus research platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Vers3Dynamics"},
]
keywords = ["ai", "research", "acoustic-physics", "multi-agent", "lm-studio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "openai>=1.0.0",
    "requests>=2.28.0",
    "numpy>=1.24.0",
    "networkx>=3.0",
    "rich>=13.0.0",
    "websockets>=12.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
search = ["duckduckgo-search>=4.0"]
scraping = ["scrapling>=0.2"]
tts = ["pyttsx3>=2.90"]
science = ["scikit-learn>=1.3.0"]
all = [
    "duckduckgo-search>=4.0",
    "scrapling>=0.2",
    "pyttsx3>=2.90",
    "scikit-learn>=1.3.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
rain-lab = "rain_lab:main"

[tool.setuptools.packages.find]
include = ["rain_lab_chat*"]

[tool.ruff]
target-version = "py310"
line-length = 120

# Exclude vendored / generated code that we don't own
exclude = [
    "rlm-main/",
    "hello_os.py",       # 53k-line Colab export — not maintained here
    ".git",
    "__pycache__",
]

[tool.ruff.lint]
# Bug detection + import ordering
select = ["E", "F", "W", "I"]
ignore = [
    "E501",   # line-length — many scripts exceed 120 today
    "E402",   # module-level import not at top — scripts do sys.path manipulation
    "F541",   # f-string without placeholders — cosmetic, not a bug
    "F841",   # local variable assigned but never used — clean up later
    "W291",   # trailing whitespace
    "W293",   # whitespace before a statement (blank lines with spaces)
    "W605",   # invalid escape sequence — raw-string Windows paths
    "E722",   # bare except — used intentionally in several scripts
]

[tool.ruff.lint.isort]
known-first-party = ["rain_lab_chat"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
exclude = [
    "rlm-main/",
    "hello_os\\.py",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
